Finding problems that need to be solved: Rendering individual parts that aren't of the Clothing class is problematic as it only takes strings and combines it with attributes of a character then looks for the corresponding path. This should be true for Face_Style (the_person.face_style and the_person.expression_images.facial_style), Tits (the_person.tits) and Body_Type (the_person.body_type).
Hair Styles (the_person.hair_style) and Skin (the_person.body_images) 
are Clothing (class) items and are easy to implement and tweak as they 
can be shown individually and do not need to be part of a Person 
(the_person).

Creating a custom / tweaked generate_item_displayable() and 
generate_emotion_displayable() might be nescessary for those parts to be 
displayable. Afterwards crop / reposition the image.